auth/tests/functional/requirements.txt
pytest==7.1.2
redis==4.3.4
faker==14.2.0
orjson==3.7.12
aioredis==1.3.1
psycopg2-binary==2.9.5
requests==2.28.1